# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ReviewFlow Pro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"

    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin/manager)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/toggle-status:
    patch:
      operationId: toggleUserStatus
      tags: [users]
      summary: Enable or disable a user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleStatusBody"
      responses:
        "200":
          description: Status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── Clients ───
...
  /qr-campaigns:
    get:
      operationId: listQrCampaigns
      tags: [qr-campaigns]
      summary: List all QR campaigns
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of QR campaigns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedQrCampaigns"

    post:
      operationId: createQrCampaign
      tags: [qr-campaigns]
      summary: Create a new QR campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQrCampaignBody"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCampaign"
  /qr-campaigns/{id}:
    get:
      operationId: getQrCampaign
      tags: [qr-campaigns]
      summary: Get QR campaign by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QR campaign found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCampaign"

    patch:
      operationId: updateQrCampaign
      tags: [qr-campaigns]
      summary: Update a QR campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQrCampaignBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCampaign"

    delete:
      operationId: deleteQrCampaign
      tags: [qr-campaigns]
      summary: Delete a QR campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /qr-campaigns/{id}/stats:
    get:
      operationId: getQrCampaignStats
      tags: [qr-campaigns]
      summary: Get analytics for a QR campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCampaignStats"

  # ─── Reviews ───
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List all reviews (positive redirects)
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: rating
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedReviews"
  /reviews/{id}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get a review by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ─── Feedback (Negative) ───
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    ToggleStatusBody:
      type: object
      properties:
        status:
          type: string
      required: [status]

    # ─── User ───
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, admin, manager, client]
        status:
          type: string
          enum: [active, inactive]
        avatar:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, email, role, status, createdAt, updatedAt]
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, manager, client]
        phone:
          type: ["string", "null"]
      required: [name, email, password, role]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        status:
          type: string
    PaginatedUsers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # ─── Client ───
...
    QrCampaign:
      type: object
      properties:
        id:
          type: integer
        qrCode:
          type: string
        name:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        destinationUrl:
          type: string
        status:
          type: string
          enum: [active, inactive, expired]
        totalScans:
          type: integer
        totalReviews:
          type: integer
        conversionRate:
          type: number
        qrImageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, qrCode, name, clientId, clientName, destinationUrl, status, totalScans, totalReviews, conversionRate, createdAt, updatedAt]
    CreateQrCampaignBody:
      type: object
      properties:
        name:
          type: string
        clientId:
          type: integer
        destinationUrl:
          type: string
      required: [name, clientId, destinationUrl]
    UpdateQrCampaignBody:
      type: object
      properties:
        name:
          type: string
        destinationUrl:
          type: string
        status:
          type: string
    PaginatedQrCampaigns:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/QrCampaign"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    QrCampaignStats:
      type: object
      properties:
        totalScans:
          type: integer
        totalReviews:
          type: integer
        positiveReviews:
          type: integer
        negativeReviews:
          type: integer
        conversionRate:
          type: number
        deviceBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/DeviceBreakdownItem"
        scansByDay:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
      required: [totalScans, totalReviews, positiveReviews, negativeReviews, conversionRate, deviceBreakdown, scansByDay]
    DeviceBreakdownItem:
      type: object
      properties:
        device:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required: [device, count, percentage]

    # ─── Review ───
    Review:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        rating:
          type: integer
        reviewText:
          type: ["string", "null"]
        keywords:
          type: array
          items:
            type: string
        redirectedToGoogle:
          type: boolean
        deviceType:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, campaignId, clientId, clientName, rating, redirectedToGoogle, keywords, createdAt]
    PaginatedReviews:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # ─── Feedback ───
...
    TimeSeriesPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: integer
        label:
          type: ["string", "null"]
      required: [date, value]
...